+++ sydtest-mutation/default.nix
@@ -1,6 +1,6 @@
 { mkDerivation, aeson, aeson-pretty, async, base, bytestring
 , containers, genvalidity-sydtest, genvalidity-sydtest-aeson, lib
-, path, path-io, sydtest, sydtest-discover
+, path, path-io, safe-coloured-text, sydtest, sydtest-discover
 , sydtest-mutation-runtime, text
 }:
 mkDerivation {
@@ -10,8 +10,8 @@ mkDerivation {
   libraryHaskellDepends = [ base sydtest sydtest-mutation-runtime ];
   testHaskellDepends = [
     aeson aeson-pretty async base bytestring containers
-    genvalidity-sydtest genvalidity-sydtest-aeson path path-io sydtest
-    sydtest-mutation-runtime text
+    genvalidity-sydtest genvalidity-sydtest-aeson path path-io
+    safe-coloured-text sydtest sydtest-mutation-runtime text
   ];
   testToolDepends = [ sydtest-discover ];
   homepage = "https://github.com/NorfairKing/sydtest#readme";